requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
pydantic==2.7.1
